ui cleanup
