+++ x-pack/plugins/synthetics/public/apps/synthetics/contexts/index.ts
@@ -9,4 +9,3 @@ export * from './synthetics_refresh_context';
 export * from './synthetics_settings_context';
 export * from './synthetics_theme_context';
 export * from './synthetics_startup_plugins_context';
-export * from './synthetics_data_view_context';
