--- x-pack/plugins/synthetics/public/apps/synthetics/state/index_status/selectors.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { createSelector } from 'reselect';
-import type { SyntheticsAppState } from '../root_reducer';
-
-const getState = (appState: SyntheticsAppState) => appState.indexStatus;
-export const selectIndexState = createSelector(getState, (state) => state);
