+++ x-pack/plugins/synthetics/public/apps/synthetics/state/root_effect.ts
@@ -18,7 +18,6 @@ import { syncGlobalParamsEffect } from './settings';
 import { fetchAgentPoliciesEffect } from './private_locations';
 import { fetchNetworkEventsEffect } from './network_events/effects';
 import { fetchSyntheticsMonitorEffect } from './monitor_details';
-import { fetchIndexStatusEffect } from './index_status';
 import { fetchSyntheticsEnablementEffect } from './synthetics_enablement';
 import {
   enableMonitorAlertEffect,
@@ -32,7 +31,6 @@ import { fetchPingStatusesEffect } from './ping_status';
 
 export const rootEffect = function* root(): Generator {
   yield all([
-    fork(fetchIndexStatusEffect),
     fork(fetchSyntheticsEnablementEffect),
     fork(upsertMonitorEffect),
     fork(fetchServiceLocationsEffect),
