+++ x-pack/plugins/synthetics/public/apps/synthetics/state/root_reducer.ts
@@ -21,7 +21,6 @@ import { agentPoliciesReducer, AgentPoliciesState } from './private_locations';
 import { networkEventsReducer, NetworkEventsState } from './network_events';
 import { monitorDetailsReducer, MonitorDetailsState } from './monitor_details';
 import { uiReducer, UiState } from './ui';
-import { indexStatusReducer, IndexStatusState } from './index_status';
 import { syntheticsEnablementReducer, SyntheticsEnablementState } from './synthetics_enablement';
 import { monitorListReducer, MonitorListState } from './monitor_list';
 import { serviceLocationsReducer, ServiceLocationsState } from './service_locations';
@@ -35,7 +34,6 @@ export interface SyntheticsAppState {
   pingStatus: PingStatusState;
   elasticsearch: QueriesState;
   monitorList: MonitorListState;
-  indexStatus: IndexStatusState;
   overview: MonitorOverviewState;
   networkEvents: NetworkEventsState;
   agentPolicies: AgentPoliciesState;
@@ -53,7 +51,6 @@ export const rootReducer = combineReducers<SyntheticsAppState>({
   settings: settingsReducer,
   pingStatus: pingStatusReducer,
   monitorList: monitorListReducer,
-  indexStatus: indexStatusReducer,
   overview: monitorOverviewReducer,
   networkEvents: networkEventsReducer,
   elasticsearch: elasticsearchReducer,
